Define ImageWarpingField
