+++ src/Core/ColorMap/ColorMapOptimization.cpp
@@ -26,66 +26,18 @@
 
 #include "ColorMapOptimization.h"
 
-#include <Core/ColorMap/ColorMapOptimizationJacobian.h>
 #include <Core/Camera/PinholeCameraTrajectory.h>
+#include <Core/ColorMap/ColorMapOptimizationJacobian.h>
+#include <Core/ColorMap/ImageWarpingField.h>
 #include <Core/Geometry/Image.h>
 #include <Core/Geometry/RGBDImage.h>
 #include <Core/Geometry/TriangleMesh.h>
 #include <Core/Utility/Console.h>
 #include <Core/Utility/Eigen.h>
-#include <IO/ClassIO/TriangleMeshIO.h>
 
 namespace open3d {
 
 namespace {
-
-class ImageWarpingField {
-
-public:
-    ImageWarpingField (int width, int height, int number_of_vertical_anchors) {
-        InitializeWarpingFields(width, height, number_of_vertical_anchors);
-    }
-
-    void InitializeWarpingFields(int width, int height,
-            int number_of_vertical_anchors) {
-        anchor_h_ = number_of_vertical_anchors;
-        anchor_step_ = double(height) / (anchor_h_ - 1);
-        anchor_w_ = int(std::ceil(double(width) / anchor_step_) + 1);
-        flow_ = Eigen::VectorXd(anchor_w_ * anchor_h_ * 2);
-        for (int i = 0; i <= (anchor_w_ - 1); i++) {
-            for (int j = 0; j <= (anchor_h_ - 1); j++) {
-                int baseidx = (i + j * anchor_w_) * 2;
-                flow_(baseidx) = i * anchor_step_;
-                flow_(baseidx + 1) = j * anchor_step_;
-            }
-        }
-    }
-
-    Eigen::Vector2d QueryFlow(int i, int j) const {
-        int baseidx = (i + j * anchor_w_) * 2;
-        // exceptional case: quried anchor index is out of pre-defined space
-        if (baseidx < 0 || baseidx > anchor_w_ * anchor_h_ * 2)
-            return Eigen::Vector2d(0.0, 0.0);
-        else
-            return Eigen::Vector2d(flow_(baseidx), flow_(baseidx + 1));
-    }
-
-    Eigen::Vector2d GetImageWarpingField(double u, double v) const {
-        int i = (int)(u / anchor_step_);
-        int j = (int)(v / anchor_step_);
-        double p = (u - i * anchor_step_) / anchor_step_;
-        double q = (v - j * anchor_step_) / anchor_step_;
-        return (1 - p) * (1 - q) * QueryFlow(i, j)
-            + (1 - p) * (q)* QueryFlow(i, j + 1)
-            + (p)* (1 - q) * QueryFlow(i + 1, j)
-            + (p)* (q)* QueryFlow(i + 1, j + 1);
-    }
-
-public:
-    Eigen::VectorXd flow_;
-    int anchor_w_, anchor_h_;
-    double anchor_step_;
-};
 
 const double IMAGE_BOUNDARY_MARGIN = 10;
