+++ src/Core/ColorMap/ImageWarpingField.h
@@ -1,0 +1,46 @@
+// ----------------------------------------------------------------------------
+// -                        Open3D: www.open3d.org                            -
+// ----------------------------------------------------------------------------
+// The MIT License (MIT)
+//
+// Copyright (c) 2018 www.open3d.org
+//
+// Permission is hereby granted, free of charge, to any person obtaining a copy
+// of this software and associated documentation files (the "Software"), to deal
+// in the Software without restriction, including without limitation the rights
+// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+// copies of the Software, and to permit persons to whom the Software is
+// furnished to do so, subject to the following conditions:
+//
+// The above copyright notice and this permission notice shall be included in
+// all copies or substantial portions of the Software.
+//
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING
+// FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS
+// IN THE SOFTWARE.
+// ----------------------------------------------------------------------------
+
+#include <Core/Geometry/Image.h>
+
+namespace open3d {
+
+class ImageWarpingField {
+
+public:
+    ImageWarpingField (int width, int height, int number_of_vertical_anchors);
+    void InitializeWarpingFields(int width, int height,
+            int number_of_vertical_anchors);
+    Eigen::Vector2d QueryFlow(int i, int j) const;
+    Eigen::Vector2d GetImageWarpingField(double u, double v) const;
+
+public:
+    Eigen::VectorXd flow_;
+    int anchor_w_, anchor_h_;
+    double anchor_step_;
+};
+
+}	// namespace open3d
